5. Na kraju programa ispisati podatke o namirnicama koja ima najveći broj kilokalorija te najveću ukupnu cijenu s obzirom na težinu.

+++ src/main/java/hr/java/production/model/Edible.java
@@ -1,9 +1,7 @@
 package hr.java.production.model;
 
-import java.math.BigDecimal;
-
 public interface Edible {
     int calculateKilocalories();
 
-    BigDecimal calculatePrice(double weight);
+    double calculatePrice(double weight);
 }
